app/api/requirements.txt
fastapi
uvicorn[standard]
azure-identity
azure-storage-blob
azure-data-tables
requests
python-multipart
